backend/requirements.txt
flask-apispec==0.11.0
Flask-SQLAlchemy==2.4.4
Flask==1.1.2
Flask-Cors==3.0.10
Flask-DB==0.3.2
Flask-Bcrypt==1.0.1
mysql-connector-python==8.0.27
marshmallow==3.14.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0
requests==2.11.1
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.38.2
pyjwt==1.4.2
Werkzeug==1.0.1
python-dotenv==0.17.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
gunicorn==20.1.0
python-dateutil==2.8.2
Flask-Mail==0.9.1
itsdangerous==2.0.1
icalendar==4.0.9
Jinja2==3.0.1
Flask-Testing==0.8.1
Flask-Migrate==3.1.0